Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -1,15 +1,16 @@
 package com.educandoweb.course.services;
 
-
-
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DataBaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 /*Anotação para registrar o UserService como componente do Spring, assim é possível injetar automaticamente com @Autowired no UserResource
@@ -20,31 +21,40 @@ import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
-	
+
 	@Autowired
 	private UserRepository repository;
-	
-	public List<User> findAll (){
+
+	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
-	public User findById (Long id){
+
+	public User findById(Long id) {
 		Optional<User> obj = repository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			
+			if (!repository.existsById(id)) {
+					throw new ResourceNotFoundException(id);
+			}
+			
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DataBaseException(e.getMessage());
+		}
 	}
-	
+
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
 		updateData(entity, obj);
-		return repository.save(entity);		
+		return repository.save(entity);
 	}
 
 	private void updateData(User entity, User obj) {
@@ -52,5 +62,5 @@ public class UserService {
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
 	}
-	
+
 }
